Fixed a bug when using multiple gamepads
If a second gamepad was connected after NESmenu was started, it would use that gamepad instead of the first one,
and of course the setup wouldn't be right and neither gamepad could control the menu.

Changed it so that when started, the highest gamepad ID is saved, and this ID is always used.
There is still a bug if the second gamepad is disconnected again, the program will crash.
I haven't figured that one out just yet...

Committer: Asmund Liseth <asmund@asmund-laptop.lan>